switch from AMD to CommonJS style
