+++ clients/shared/backdrop.js
@@ -1,0 +1,18 @@
+
+function Backdrop() {
+	this.root = document.getElementById('backdrop');
+	this.text = this.root.querySelector('.bdp-text');
+	this.subtext = this.root.querySelector('.bdp-subtext');
+}
+
+Backdrop.prototype.update = function (text, subtext, visible) {
+	this.text.textContent = text;
+	this.subtext.textContent = subtext;
+	this.root.classList.toggle('hidden', !visible);
+};
+
+Backdrop.prototype.hide = function () {
+	this.root.classList.add('hidden');
+};
+
+module.exports.Backdrop = Backdrop;
