+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "author": "Axel Bocciarelli",
   "description": "Scoring software for ITF Taekwon-Do sparring matches",
+  "license": "SEE LICENSE IN LICENSE.md",
   "repository": {
     "type": "git",
     "url": "http://github.com/axelboc/taekwon"
@@ -19,22 +20,22 @@
   "dependencies": {
     "assert-dotenv": "2.0.*",
     "async": "1.2.*",
-    "connect-nedb-session": "0.0.*",
-    "cookie": "0.1.*",
-    "cookie-parser": "1.3.*",
     "express": "4.12.*",
     "express-handlebars": "2.0.*",
-    "express-session": "1.11.*",
+    "fastclick": "^1.0.6",
+    "handlebars": "^3.0.3",
     "javascript-state-machine": "2.3.*",
     "nedb": "1.1.*",
     "nedb-logger": "0.1.*",
     "primus": "3.1.*",
     "primus-emit": "0.1.*",
-    "sockjs": "0.3.*"
+    "sockjs": "0.3.*",
+    "tiny-cookie": "^0.5.4"
   },
   "devDependencies": {
     "browserify": "^10.2.4",
     "chai": "3.0.*",
+    "gulp": "^3.9.0",
     "mocha": "2.2.*",
     "sinon": "1.15.*"
   }
