+++ templates/jury-president.hbs
